script/requirements.txt
pyautogui
pynput
pyperclip
pydirectinput
colorama
pytesseract 
Pillow
keyboard
opencv-python 
numpy
requests
curl_cffi
certifi
openai 
httpx
httpx[socks]
pandas